Final Calorie Counter Backend Update
At least, hopefully.
The calorie counter is fully functional!
Can check different days and successfully read what was logged on that date. It also redirects you to the login page if you try to view it without an account.

I changed the schema completely, literally deleted 3 tables.

Also, I commented literally everything so this should hopefully be easy to follow along with.